requirements.txt
# ============================================================
# DEPENDENCIAS DEL AGENTE INTELIGENTE - Python 3.12+
# ============================================================

# LangChain y integración con Groq (dejar que pip resuelva compatibilidad)
langchain>=0.2.0
langchain-groq>=0.1.0

# Groq API Client
groq>=0.9.0

# FastAPI y Pydantic (compatibles)
fastapi>=0.100.0
uvicorn>=0.25.0
pydantic>=2.5.0

# Data processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0

# Utilidades
python-dotenv>=1.0.0
python-json-logger>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
